Improved the solution files to not include the dll lib-files when building. Implemented basic command line options for StormMain and added support for parameters in buildconfig.

+++ StormGui/StormGui.cpp
@@ -26,7 +26,8 @@ namespace storm {
 	LibData::LibData() {}
 
 	LibData::~LibData() {
-		app->terminate();
+		if (app)
+			app->terminate();
 	}
 
 }
